+++ projections/typescript/index.ts
@@ -1,2 +1,3 @@
 export * from "./capsule.js";
+export * from "./validator-routes.js";
 export * from "./validator-api.js";